requirements.txt
aiofiles==23.2.1
altair==5.2.0
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
appdirs==1.4.4
argcomplete==3.5.0
asgiref==3.8.1
attrs==23.2.0
audiocraft==1.2.0
audioread==3.0.1
av==12.0.0
backoff==2.2.1
blis==0.7.11
catalogue==2.0.10
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
cloudpickle==3.0.0
colorama==0.4.6
colorlog==6.8.2
comtypes==1.4.6
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
decorator==5.1.1
DeepFilterLib==0.5.6
DeepFilterNet==0.5.6
demucs==4.0.1
distlib==0.3.8
Django==3.2
django-cors-headers==4.4.0
djangorestframework==3.12.4
docopt==0.6.2
docstring_parser==0.16
dora_search==0.1.12
einops==0.7.0
encodec==0.1.1
# Editable install with no version control (fam==0.1.0)
-e c:\users\richard\documents\python\codeignite\tts\metavoice-src
fastapi==0.110.0
ffmpy==0.3.2
filelock==3.13.3
flashy==0.0.2
fonttools==4.50.0
fsspec==2024.3.1
gradio==4.24.0
gradio_client==0.14.0
gTTS==2.5.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.21.4
hydra-colorlog==1.2.0
hydra-core==1.3.2
identify==2.5.35
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
julius==0.2.7
kiwisolver==1.4.5
lameenc==1.7.0
langcodes==3.3.0
lazy_loader==0.3
librosa==0.10.1
lightning-utilities==0.11.2
llvmlite==0.42.0
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
monotonic==1.6
mpmath==1.3.0
msgpack==1.0.8
murmurhash==1.0.10
mysqlclient==2.2.4
networkx==3.2.1
ninja==1.11.1.1
nodeenv==1.8.0
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
omegaconf==2.3.0
openunmix==1.2.1
orjson==3.10.0
packaging==23.2
pandas==2.2.1
pillow==10.2.0
pipx==1.6.0
platformdirs==4.2.0
pluggy==1.4.0
pooch==1.8.1
posthog==3.5.0
pre-commit==3.7.0
preshed==3.0.9
protobuf==4.25.3
PyAudio==0.2.14
pycparser==2.21
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
Pygments==2.17.2
PyMySQL==1.1.1
pyparsing==3.1.2
pypiwin32==223
pytest==8.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.9
pyttsx3==2.91
pytz==2024.1
pywin32==306
PyYAML==6.0.1
referencing==0.34.0
regex==2023.12.25
requests==2.32.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.4
safetensors==0.4.2
scikit-learn==1.4.1.post1
scipy==1.12.0
semantic-version==2.10.0
sentencepiece==0.2.0
shellingham==1.5.4
shtab==1.7.1
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
sqlparse==0.5.1
srsly==2.4.8
starlette==0.36.3
submitit==1.5.1
sympy==1.12
thinc==8.2.3
threadpoolctl==3.4.0
tiktoken==0.5.1
tokenizers==0.15.2
tomlkit==0.12.0
toolz==0.12.1
torch==2.2.1
torchaudio==2.2.1
torchmetrics==1.3.2
tqdm==4.66.2
transformers==4.39.2
treetable==0.2.5
typer==0.9.4
typing_extensions==4.10.0
tyro==0.7.3
tzdata==2024.1
urllib3==2.2.2
userpath==1.9.2
uvicorn==0.27.1
views==0.3
virtualenv==20.25.1
wasabi==1.1.2
weasel==0.3.4
websockets==11.0.3
win32-setctime==1.1.0
xformers==0.0.27.post2+cu118
